Fix spelling mistake of 'utils.py'
